test: move tests/engine to tests/core and tests/analysis matching real subjects
