+++ tests/analysis/test_adaptive_scaling_integration.py
