+++ tests/analysis/test_consolidated_integration.py
