+++ tests/analysis/test_manifest.py
