+++ tests/analysis/test_subagent_pool_merge.py
