+++ tests/analysis/test_universal_detection.py
