+++ tests/core/test_evidence_parser_line.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import json
 
 from quodeq.core.evidence.parser import _parse_jsonl_line
-from tests.engine.conftest import _evidence_line
+from tests._evidence_helpers import _evidence_line
 
 
 class TestParseJsonlLine:
