+++ tests/core/test_finding_builder.py
