+++ tests/core/test_mcp_refs_multi.py
