+++ tests/core/test_scoring_internals.py
