+++ tests/core/test_scoring_params.py
