+++ tests/core/test_standards.py
