--- tests/engine/fixtures/sample_src/bad.ts
@@ -1,1 +1,0 @@
-const result = eval(userInput);
